+++ src/strictenv/__init__.py
@@ -12,13 +12,14 @@ from .errors import (
 )
 from .fields import Field
 from .structs import TransformStruct
-from .transforms import transform
+from .transforms import transform, transform_struct
 
 __all__ = [
     "BaseSettings",
     "TransformStruct",
     "Field",
     "transform",
+    "transform_struct",
     "SettingsError",
     "EnvFileNotFoundError",
     "EnvFileReadError",